Use the newest version of gocd agents job_run_history api.
The newest version of https://api.gocd.org/20.3.0/#agent-job-run-history
has changed quite a bit. The structure is now different to the regular
job history api which means that we need to create new data structures
specific to agent job history.

There are a few new parameters tot the agent-job-history-api, of which
offset and page_size are two of them. They have been introduced in the
signature of AgentRunJobHistory as well. sort_column and sort_order has
been ignored for now but can be added when there is a need.